requirements.txt
openai-whisper
GPUtil
numpy
sounddevice
librosa
soundfile
pydub
ffmpeg-python
